Clean up qr feed
