+++ src/renderer/components/Feed/Feeds/CameraFeed.tsx
@@ -8,6 +8,7 @@ import { useActor } from '@xstate/react'
 import * as React from 'react'
 import { FC, useEffect, useRef } from 'react'
 import { log } from '@/renderer/logger'
+import { QRFeed } from './QRFeed/QRFeed'
 
 interface Props {
   feed: ICameraFeed
@@ -88,6 +89,7 @@ const Webcam: FC<{ deviceid: string; flipped: boolean; rotated: boolean }> = ({
 
 const View: FC<Props> = ({ feed }) => {
   const source = useSelector(selectVideoUrl(feed.camera))
+  const imageRef = useRef<HTMLImageElement | null>(null)
 
   switch (feed.camera.type) {
     case CameraType.COMPRESSED:
@@ -118,6 +120,18 @@ const View: FC<Props> = ({ feed }) => {
           flipped={feed.camera.flipped}
           rotated={feed.camera.rotated}
         />
+      )
+    case CameraType.QR_CODE:
+      return (
+        <QRFeed imageRef={imageRef}>
+          <StyledImg
+            src={source}
+            flipped={feed.camera.flipped}
+            rotated={feed.camera.rotated}
+            ref={imageRef}
+            alt="camera stream"
+          />
+        </QRFeed>
       )
     default:
       return <TextFeed text="stream type not supported" />
